remove TAGGED logic

+++ libs/base/buffer.cpp
@@ -203,7 +203,7 @@ static int writeBytes(uint8_t *dst, uint8_t *src, int length, bool swapBytes, in
     } else {
         if (length == 4 && ((uint32_t)dst & 3) == 0)
             *(uint32_t *)dst = *(uint32_t *)src;
-        else if (length == 2 && ((uint16_t)dst & 1) == 0)
+        else if (length == 2 && ((uint32_t)dst & 1) == 0)
             *(uint16_t *)dst = *(uint16_t *)src;
         else
             memcpy(dst, src, length);
@@ -225,7 +225,7 @@ static int readBytes(uint8_t *src, uint8_t *dst, int length, bool swapBytes, int
     } else {
         if (length == 4 && ((uint32_t)src & 3) == 0)
             *(uint32_t *)dst = *(uint32_t *)src;
-        else if (length == 2 && ((uint16_t)src & 1) == 0)
+        else if (length == 2 && ((uint32_t)src & 1) == 0)
             *(uint16_t *)dst = *(uint16_t *)src;
         else
             memcpy(dst, src, length);
